+++ Lib/fontbakery/profiles/fontbureau.py
@@ -10,6 +10,7 @@ PROFILE = {
         "colorfont_tables",
         "empty_glyph_on_gid1_for_colrv0",
         "empty_letters",
+        "epar",
         "family/control_chars",
         "file_size",
         "fvar_name_entries",