+++ Lib/fontbakery/profiles/universal.py
@@ -34,6 +34,7 @@ PROFILE = {
             "contour_count",
             "empty_glyph_on_gid1_for_colrv0",
             "empty_letters",
+            "epar",
             "family/control_chars",
             "family/single_directory",
             "family/vertical_metrics",